+++ include/Zeni/Rete/Token_Index.hpp
@@ -12,7 +12,7 @@ namespace Zeni::Rete {
   class ZENI_RETE_LINKAGE Token_Index {
   public:
     Token_Index();
-    Token_Index(const int64_t rete_row_, const int64_t token_row_, const int8_t column_);
+    Token_Index(const int64_t rete_row_, const int64_t token_row_, const int64_t column_);
 
     bool operator==(const Token_Index &rhs) const;
     bool operator!=(const Token_Index &rhs) const;
@@ -23,7 +23,7 @@ namespace Zeni::Rete {
 
     int64_t rete_row;
     int64_t token_row;
-    int8_t column;
+    int64_t column;
   };
 
 }
