+++ include/Zeni/Rete/Variable_Indices.hpp
@@ -29,11 +29,11 @@ namespace Zeni::Rete {
     ZENI_RETE_LINKAGE virtual Token_Index find_index(const std::string &name) const = 0;
 
     ZENI_RETE_LINKAGE virtual std::string_view find_name(const Token_Index &index) const = 0;
-    ZENI_RETE_LINKAGE virtual std::string_view find_name_rete(const int64_t rete_row, const int8_t column) const = 0;
-    ZENI_RETE_LINKAGE virtual std::string_view find_name_token(const int64_t token_row, const int8_t column) const = 0;
+    ZENI_RETE_LINKAGE virtual std::string_view find_name_rete(const int64_t rete_row, const int64_t column) const = 0;
+    ZENI_RETE_LINKAGE virtual std::string_view find_name_token(const int64_t token_row, const int64_t column) const = 0;
 
-    ZENI_RETE_LINKAGE virtual std::pair<std::string_view, Token_Index> find_both_rete(const int64_t rete_row, const int8_t column) const = 0;
-    ZENI_RETE_LINKAGE virtual std::pair<std::string_view, Token_Index> find_both_token(const int64_t token_row, const int8_t column) const = 0;
+    ZENI_RETE_LINKAGE virtual std::pair<std::string_view, Token_Index> find_both_rete(const int64_t rete_row, const int64_t column) const = 0;
+    ZENI_RETE_LINKAGE virtual std::pair<std::string_view, Token_Index> find_both_token(const int64_t token_row, const int64_t column) const = 0;
 
     ZENI_RETE_LINKAGE virtual void insert(const std::string_view name, const Token_Index &index) = 0;
